KEYBOARD_GENERATOR/src/Testers/QAPAlgorithmTest.java
package Testers;

import static org.junit.Assert.*;

import Domain.QAPAlgorithm;
import org.junit.Test;

/**
 * Test class for the QAPAlgorithm class.
 */
public class QAPAlgorithmTest {

    /**
     * Test the {@code getPositions} method of the QAPAlgorithm class.
     */
    @Test
    public void testGetPositions() {
        // Example test data, modify as needed
        int[][] transitTable = {
                {0, 1, 2},
                {1, 0, 3},
                {2, 3, 0}
        };

        int[][] greedyPositions = {
                {0, 1, 2},
                {1, 2, 0},
                {2, 0, 1}
        };

        int initialBound = 10;

        // Creating a QAPAlgorithm object
        QAPAlgorithm qapAlgorithm = new QAPAlgorithm();

        // Calling the getPositions method and getting the result
        int[][] result = qapAlgorithm.getPositions(transitTable, greedyPositions, initialBound);

        // Assert that the result is not null or empty, and add more specific assertions based on your requirements
        assertNotNull("Result should not be null.", result);
        assertTrue("Result should have at least one row.", result.length > 0);
        assertTrue("Result should have at least one column.", result[0].length > 0);
    }
}